Move components into their own folder
