+++ ui/src/components/wish/Wish.test.js
